src/circuits/qaoa.py
# src/circuits/qaoa.py

from qiskit import QuantumCircuit, execute, Aer
import numpy as np
from typing import List, Dict, Tuple, Any

class QAOA:
    """
    Implementation of the Quantum Approximate Optimization Algorithm (QAOA)
    for solving combinatorial optimization problems.
    """
    
    def __init__(self, n_qubits: int):
        """
        Initialize QAOA solver.
        
        Args:
            n_qubits: Number of qubits in the system
        """
        self.n_qubits = n_qubits
        
    def create_qaoa_circuit(self, 
                          graph_edges: List[Tuple[int, int, float]], 
                          gamma: float, 
                          beta: float) -> QuantumCircuit:
        """
        Create a QAOA circuit for the MaxCut problem.
        
        Args:
            graph_edges: List of (node1, node2, weight) tuples representing graph edges
            gamma: Mixing angle parameter
            beta: Problem hamiltonian angle parameter
            
        Returns:
            QuantumCircuit: QAOA circuit implementation
        """
        qc = QuantumCircuit(self.n_qubits, self.n_qubits)
        
        # Initial state superposition
        qc.h(range(self.n_qubits))
        qc.barrier()
        
        # Problem unitary
        for edge in graph_edges:
            k, l, weight = edge
            qc.cu(0, 0, -2*gamma*weight, 0, k, l)
            qc.u(0, 0, gamma*weight, k)
            qc.u(0, 0, gamma*weight, l)
            
        # Mixing unitary
        qc.barrier()
        qc.rx(2*beta, range(self.n_qubits))
        
        # Measurement
        qc.measure(range(self.n_qubits), range(self.n_qubits))
        
        return qc
        
    def execute_circuit(self, 
                       circuit: QuantumCircuit, 
                       shots: int = 1000) -> Dict[str, int]:
        """
        Execute QAOA circuit and return measurement results.
        
        Args:
            circuit: QAOA circuit to execute
            shots: Number of shots for the simulation
            
        Returns:
            Dict of measured states and their counts
        """
        simulator = Aer.get_backend('qasm_simulator')
        result = execute(circuit, simulator, shots=shots).result()
        return result.get_counts(circuit)
        
    def compute_expectation(self, 
                          counts: Dict[str, int], 
                          graph_edges: List[Tuple[int, int, float]]) -> float:
        """
        Compute expectation value of the cost Hamiltonian.
        
        Args:
            counts: Measurement results from circuit execution
            graph_edges: Graph edges with weights
            
        Returns:
            float: Expected value of the cost Hamiltonian
        """
        total_shots = sum(counts.values())
        expectation = 0
        
        for bitstring, count in counts.items():
            cost = 0
            for i, j, weight in graph_edges:
                if bitstring[i] != bitstring[j]:
                    cost += weight
            expectation += cost * count / total_shots
            
        return expectation